+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Upcoming Changes (unreleased)
 - fixing to override default country when + passed in the beginning of phone
+- fixed error when int passed as phone number
+- fixed error when ```";"``` passed as phone number
+
 
 ## 0.6.2 - June 14, 2016
 - fixed bug in ```international``` method when no country can be defined